+++ instrumentation/opentelemetry-instrumentation-aiohttp-client/src/opentelemetry/instrumentation/aiohttp_client/__init__.py
@@ -180,26 +180,23 @@ def create_trace_config(
 
         http_method = params.method.upper()
         request_span_name = f"HTTP {http_method}"
+        request_url = (
+            remove_url_credentials(trace_config_ctx.url_filter(params.url))
+            if callable(trace_config_ctx.url_filter)
+            else remove_url_credentials(str(params.url))
+        )
+
+        span_attributes = {
+            SpanAttributes.HTTP_METHOD: http_method,
+            SpanAttributes.HTTP_URL: request_url,
+        }
 
         trace_config_ctx.span = trace_config_ctx.tracer.start_span(
-            request_span_name,
-            kind=SpanKind.CLIENT,
+            request_span_name, kind=SpanKind.CLIENT, attributes=span_attributes
         )
 
         if callable(request_hook):
             request_hook(trace_config_ctx.span, params)
-
-        if trace_config_ctx.span.is_recording():
-            attributes = {
-                SpanAttributes.HTTP_METHOD: http_method,
-                SpanAttributes.HTTP_URL: remove_url_credentials(
-                    trace_config_ctx.url_filter(params.url)
-                )
-                if callable(trace_config_ctx.url_filter)
-                else remove_url_credentials(str(params.url)),
-            }
-            for key, value in attributes.items():
-                trace_config_ctx.span.set_attribute(key, value)
 
         trace_config_ctx.token = context_api.attach(
             trace.set_span_in_context(trace_config_ctx.span)